finish adding and removing from cart

+++ backend/db.py
@@ -58,21 +58,40 @@ def add_product_to_cart(user_id, product_id, quantity):
    
     if existing_product:
         new_quantity = existing_product[1] + quantity
-        c.execute("UPDATE cartcontent SET quantity = ? WHERE id = ?;", (new_quantity, existing_product[0]))
+        c.execute("UPDATE cartcontent SET quantity =" + str(new_quantity)+ " WHERE card_id = ? and product_id = ?;", (user_id, product_id))
     else:
         c.execute("INSERT INTO cartcontent (card_id, product_id, quantity) VALUES (?, ?, ?);", (user_id, product_id, quantity))
     
 
     c.execute("SELECT total_price, money_saved FROM cart WHERE user_id = ?;", (user_id,))
     curr_totals = c.fetchall()
-    c.execute("UPDATE cart SET total_price = ? WHERE user_id = ?;", (float(curr_totals[0][0])+modified_product_price, user_id))
-    c.execute("UPDATE cart SET money_saved = ? WHERE user_id = ?;", (float(curr_totals[0][1])+float(product_price[0]), user_id))
+    c.execute("UPDATE cart SET total_price = ? WHERE user_id = ?;", (round(float(curr_totals[0][0]), 3)+modified_product_price, user_id))
+    c.execute("UPDATE cart SET money_saved = ? WHERE user_id = ?;", (round(float(curr_totals[0][1]), 3)+round(float(product_price[0]), 3), user_id))
 
     db.commit()
 
 def remove_product_from_cart(user_id, product_id):
     c = db.cursor()
-    c.execute("DELETE FROM cartcontent WHERE card_id = ? AND product_id = ?;", (user_id, product_id))
+    #check if there are multiple of product
+    c.execute("SELECT id, quantity FROM cartcontent WHERE quantity > 1 AND card_id = ? AND product_id = ?;", (user_id, product_id))
+    multiple = c.fetchone()
+
+    if multiple:
+        new_quantity = multiple[1] - 1
+        c.execute("UPDATE cartcontent SET quantity =" + str(new_quantity)+ " WHERE card_id = ? and product_id = ?;", (user_id, product_id))
+    else:
+        c.execute("DELETE FROM cartcontent WHERE card_id = ? AND product_id = ?;", (user_id, product_id))
+
+    c.execute("SELECT price FROM products where id = ?;", (product_id,))
+    product_price = c.fetchone()
+
+    c.execute("SELECT modifier FROM users WHERE id = ?;", [user_id])
+    modified_product_price  = round(c.fetchone()[0], 3) * round(float(product_price[0]), 3)
+
+    c.execute("SELECT total_price, money_saved FROM cart WHERE user_id = ?;", (user_id,))
+    curr_totals = c.fetchall()
+    c.execute("UPDATE cart SET total_price = ? WHERE user_id = ?;", (round(float(curr_totals[0][0]), 3)-round(modified_product_price, 3), user_id))
+    c.execute("UPDATE cart SET money_saved = ? WHERE user_id = ?;", (round(float(curr_totals[0][1]), 3)-round(float(product_price[0]), 3), user_id))
     db.commit()
 
 def empty_cart(user_id):
@@ -117,10 +136,15 @@ def add_product(product_id, product_price, product_name, product_sku, prod_categ
 
 
 
-"""
+
+
+create_tables()
+
 c.execute("DELETE FROM products")
 c.execute("DELETE FROM users")
 c.execute("DELETE FROM cart")
+c.execute("DELETE FROM cartcontent")
+
 create_tables()
 populate_tables()
 
@@ -138,5 +162,9 @@ for prod in prod_rows:
     print(prod)
 
 add_product_to_cart("512380", '1', 1)
+add_product_to_cart("512380", '1', 1)
+
 print(get_cart("512380"))
-"""
+remove_product_from_cart("512380", '1')
+print(get_cart("512380"))
+empty_cart("512380")
